Add support for powernv builds to qemu tests
A kernel built with the powernv defconfig should be tested on an
emulated powernv machine. To do this with qemu, the `-machine`
parameter must be given a powernv value that aligns with the cpu type.
For now the only two options are `powernv8` and `powernv9` for `power8`
and `power9` respectively. Other test conditions must also be adjusted,
the machine type should look for powernv insted of pseries. To address
these differences, a `qemu-powernv` script has been created, it closely
resembles `qemu-pseries` but makes the appropriate adjustments for the
machine differences.